add third driveMode
Since RobotMap.driveMode was a boolean, only two drive modes were possible. This changes it so we can have as many as we want.

The way it is set up, 0 corresponds to arcade drive with the controller, 1 responds to tank drive with the controller, and 2 responds to arcade drive with the joystick (using the rotation axis of the joystick as turning).