fix the goto command #2
Prior to this, the values of the goto command were calculated in the parsing section of the autonomous. So the point from which the values were calculated was always [0, 0]. 

The calculation was thus moved to the DriveAutonomous initialize function. 

Note that this impementation is not the final one. The DriveAutonomous function desperately needs refactoring, now that the code is (presumably) true.