change CollectorIntake command
Since we will need to intake for an uncertain period of time, we will make the CollectorIntake take either 0 or 1. If it is 0, it will intake on hold speed. If it is 1, it will intake on full speed.

This command is designed to be interrupted by CollectorOuttake. After that, the collector won't need to be running anymore.