remove FourbarRaise and FourbarHold, add FourbarManualControl
It turns out that the FourbarRaise and Hold was not that good of an idea.

Now, the same approach as to the collector is adopted: have it in the manualControl function (and save, whether we want to hold or not).

And also add the forgotten variable changes for the Collector (oops).