update elevator movement
The movement of the elevator no longer shifts the polynomial function depending on the start and end of the ElevatorToHeight command.

Also, the elevator now goes full speed up and full speed down until getting close to the goal.

Lastly, the encoder value on the elevator resets itself if the elevator is below the elevator hold threshold, and the encoder value is repeated multiple times (the elevator is not moving).